Worked on entities move() method.
Doesn't reflect the real implementation yet, but it's a start. We have to decide whether we should use delta time or game ticks to update and render each frame. Right now, I'm using dummy values as delta times. This stuff sort of relies on the underlying game engine.

It's tricky to write code for this now, since entities will move a certain distance based on how long the player presses and holds the direction keys.