Made the GameModel listen to navigation events.
We didn't want the NavigationController to manipulate the player directly. Thus, we implemented a messaging service in order to let the navigation send arbitrary messages to the listener (GameModel) which manipulated its entities (the player) depending on the event property name.

This enables the NavigationController to solely focus on handle keyboard events, and "translate" them to directions and/or other action events such as "Fire" or "Show weapon menu".

The GameModel is free to manipulate its objects according to the events.