Tile sorting, autoincrementing tile type, scrollable list.
Added functionality to automatically increment the tile type
so that this does not need to be entered explicitly. A menu
option has been added to set the current tile type value
to start incrementing from.

Added sorting functionality.

Put the list inside a JScrollPane to make it possible to scroll
down the list.

Signed-off-by: John Barbero Unenge <john.barbero.unenge@gmail.com>