Refactored collision detection and effects of collisions
- Entities now collides correctly with walls, amongst others
- Introduced new pattern for effects after collisions. CollidableObject
now contains an abstract callback 'didCollide' which takes another object. This
callback is thus available in all collidables, and is called on collision.

This makes the code more modular, since the effect (if there is one) is delegated
to the relevant objects, and we avoid superfluous 'if' blocks in the game logic loop
in GameModel.

All in all, prettier code :)

/ By Jesper and Johan