Added status hud.
Added an actor for the status hud and added functionality to
display and update it properly.

Moved the GLBitmapFont that showed the amount of ammo on top
of the status hud.

Added another lable, scoreMeter, which will be used score rather
than the wave count (which is what we currently have).

Signed-off-by: John Barbero Unenge <john.barbero.unenge@gmail.com>